+++ apollo-router/src/plugins/demand_control/fixtures/measure_on_execution_response.router.yaml
@@ -3,5 +3,5 @@ experimental_demand_control:
   mode: measure
   strategy:
     test:
-      stage: on_execution_response
+      stage: execution_response
       error: estimated_cost_too_expensive
